API integration to Home screen -Remove button "Ir a home" -Remove button "Get Top List" -Remove function in login.js and move to /spotify/loadData.js -Add function in Login Button of Facebook and log in with email and password.
