+++ src/redux/Actions.js
@@ -2,48 +2,43 @@ export const SET_NAME = 'SET_NAME';
 export const SET_EMAIL = 'SET_EMAIL';
 export const SET_TOKEN = 'SET_TOKEN';
 export const SET_PASSWORD = 'SET_PASSWORD';
-
 export const SET_NEWPASSWORD = 'SET_NEWPASSWORD';
 export const SET_PROVIDER = 'SET_PROVIDER';
-export const SET_NEWEMAIL = 'SET_NEWEMAIL'
-
-export const setName = name => ({
-
+export const SET_NEWEMAIL = 'SET_NEWEMAIL';
 export const SET_INFO = 'SET_INFO';
-export const SET_TOPLIST = 'SET_TOPLIST'
+export const SET_TOPLIST = 'SET_TOPLIST';
 
 export const setName = (name) => ({
-
   type: SET_NAME,
   payload: name,
 });
-export const setEmail = email => ({
+export const setEmail = (email) => ({
   type: SET_EMAIL,
   payload: email,
 });
-export const setToken = token => ({
+export const setToken = (token) => ({
   type: SET_TOKEN,
   payload: token,
 });
-export const setPassword = password => ({
+export const setPassword = (password) => ({
   type: SET_PASSWORD,
   payload: password,
 });
 
-
-export const setNewPassword = newPassowrd => ({
+export const setNewPassword = (newPassowrd) => ({
   type: SET_NEWPASSWORD,
   payload: newPassowrd,
 });
 
-export const setNewEmail = newEmail => ({
+export const setNewEmail = (newEmail) => ({
   type: SET_NEWEMAIL,
-  payload: newEmail
-})
+  payload: newEmail,
+});
 
-export const setProvider = provider => ({
+export const setProvider = (provider) => ({
   type: SET_PROVIDER,
-  payload: provider
+  payload: provider,
+});
 
 export const setinfo = (info) => ({
   type: SET_INFO,
@@ -52,6 +47,5 @@ export const setinfo = (info) => ({
 
 export const setTopList = (topList) => ({
   type: SET_TOPLIST,
-  payload: topList
-
-})
+  payload: topList,
+});
