+++ src/screens/Login.js
@@ -3,19 +3,16 @@ import {Container, Boton, Logo, Texto} from '../assets/styled.js';
 import {useNavigation} from '@react-navigation/native';
 import {OR, InputLog, logInUser, CurrentProfile} from '../components/index';
 import {Store} from '../redux/Store';
-
 import {
   setEmail,
   setNewPassword,
   setPassword,
   setNewEmail,
-  setToken,
-   setinfo
-} from '../redux/Actions.js';
+} from '../redux/Actions';
 import {LogInButton} from '../components/FaceBook Button/LogInButton.js';
 import auth from '@react-native-firebase/auth';
-import {getToken, getCategories, getToplist} from '../assets/spotify/spotify_token.js';
-
+import {getToken} from '../spotify/spotify_token';
+import {loadData} from '../spotify/loadData';
 
 export const Login = () => {
   const navigation = useNavigation();
@@ -30,43 +27,16 @@ export const Login = () => {
           index: 0,
           routes: [{name: 'Main'}],
         });
-
         CurrentProfile();
       }
-
-      }
-      getToken().catch(e => console.log('first', e));
-
     });
+    getToken().catch(e => console.log('ERROR DE TOKEN', e));
+
     return subscribe;
   }, []);
 
   return (
     <Container>
-      <Boton
-        onPress={async () => {
-          await getCategories(
-            Store.getState().token,
-            'https://api.spotify.com/v1/browse/categories',
-          )
-            .then(() => {
-              getToplist(
-                Store.getState().token,
-                'https://api.spotify.com/v1/playlists/37i9dQZEVXbO3qyFxbkOE1/tracks?offset=0&limit=3',
-              )
-            })
-            .finally(()=>{
-              setTimeout(function () {
-                navigation.navigate('Main');
-              }, 1000);
-            })
-            .catch(err => {
-              console.log(err);
-            });
-        }}>
-        <Texto style={{color: 'black'}}>Ir a home</Texto>
-      </Boton>
-
       <Logo />
 
       <InputLog
@@ -87,19 +57,16 @@ export const Login = () => {
         onPress={() => {
           setHidePassword(prevState => !prevState);
         }}
-
         onChangeText={valor => {
           Store.dispatch(setPassword(valor));
           Store.dispatch(setNewPassword(valor));
         }}
-
-        onChangeText={value => {}}
-
       />
 
       <Boton
         onPress={() => {
           logInUser();
+          loadData();
         }}>
         <Texto style={{color: 'black'}}>INICIAR SESION</Texto>
       </Boton>
@@ -117,4 +84,3 @@ export const Login = () => {
     </Container>
   );
 };
-
