+++ src/spotify/spotify_token.js
@@ -2,8 +2,8 @@
 import React, {useState, useEffect} from 'react';
 import base64 from 'react-native-base64';
 import axios from 'axios';
-import { Store } from '../../redux/Store';
-import { setinfo, setToken, setTopList } from '../../redux/Actions';
+import { Store } from '../redux/Store';
+import { setinfo, setToken, setTopList } from '../redux/Actions';
 const apiPrefix = 'https://accounts.spotify.com/api';
 const client_id = '5914e5016a704b0c84b27239cfee6242';
 const client_secret = '02a63d27435a4e85a2f1e84048657e18';
@@ -21,13 +21,16 @@ export const getToken = async () => {
       },
       data: 'grant_type=client_credentials',
     }).then(tokenresponse => {
+      
       console.log(tokenresponse.data.access_token)
+      
       Store.dispatch(setToken(tokenresponse.data.access_token))
-      return tokenresponse.data.access_token 
+
+      return ; 
+
     });
   } catch (error) {
-    console.log(error);
-    return console.log('return del error')
+    console.log('ERROR DE TOKEN ' +error);
   }
 };
 
@@ -41,7 +44,7 @@ export const getCategories = async (token, uri) => {
         Authorization: 'Bearer ' + token,
       },
     })
-      .then(trackresponse => {     
+      .then(trackresponse => {            
         Store.dispatch(setinfo(trackresponse.data.categories.items))
         return trackresponse.data.categories.items;
       })
